Repalce finetune_lm.py with new finetune_lm.py using Estimator
