+++ scripts/sentiment_analysis/finetune_lm.py
@@ -1,7 +1,6 @@
 """
 Fine-tune Language Model for Sentiment Analysis
 ===============================================
-
 This example shows how to load a language model pre-trained on wikitext-2 in Gluon NLP Toolkit model
 zoo, and reuse the language model encoder for sentiment analysis on IMDB movie reviews dataset.
 """
@@ -23,19 +22,21 @@ zoo, and reuse the language model encoder for sentiment analysis on IMDB movie r
 # specific language governing permissions and limitations
 # under the License.
 
+
 import argparse
 import time
 import random
-import glob
 import multiprocessing as mp
-
 import numpy as np
+import warnings
 
 import mxnet as mx
-from mxnet import gluon, autograd
+from mxnet import gluon, autograd, metric
 from mxnet.gluon import HybridBlock
 from mxnet.gluon.data import DataLoader
-
+from mxnet.gluon.contrib.estimator import Estimator
+from mxnet.gluon.contrib.estimator.event_handler import EarlyStoppingHandler, CheckpointHandler
+from mxnet.gluon.contrib.estimator.batch_processor import BatchProcessor
 import gluonnlp as nlp
 
 nlp.utils.check_version('0.7.0')
@@ -47,44 +48,79 @@ mx.random.seed(10000)
 tokenizer = nlp.data.SpacyTokenizer('en')
 length_clip = nlp.data.ClipSequence(500)
 
-
-parser = argparse.ArgumentParser(description='MXNet Sentiment Analysis Example on IMDB. '
-                                             'We load a LSTM model that is pre-trained on '
-                                             'WikiText as our encoder.')
-parser.add_argument('--lm_model', type=str, default='standard_lstm_lm_200',
-                    help='type of the pre-trained model to load, can be "standard_lstm_200", '
-                         '"standard_lstm_200", etc.')
-parser.add_argument('--use-mean-pool', type=bool, default=True,
-                    help='whether to use mean pooling to aggregate the states from '
-                         'different timestamps.')
-parser.add_argument('--no_pretrained', action='store_true',
+parser = argparse.ArgumentParser(
+    description='MXNet Sentiment Analysis Example on IMDB. '
+                'We load a LSTM model that is pre-trained on '
+                'WikiText as our encoder.')
+parser.add_argument(
+    '--lm_model',
+    type=str,
+    default='standard_lstm_lm_200',
+    help='type of the pre-trained model to load, can be "standard_lstm_200", '
+         '"standard_lstm_200", etc.')
+parser.add_argument(
+    '--use-mean-pool',
+    type=bool,
+    default=True,
+    help='whether to use mean pooling to aggregate the states from '
+         'different timestamps.')
+parser.add_argument('--no_pretrained',
+                    action='store_true',
                     help='Turn on the option to just use the structure and '
                          'not load the pre-trained weights.')
-parser.add_argument('--lr', type=float, default=2.5E-3,
+parser.add_argument('--lr',
+                    type=float,
+                    default=2.5E-3,
                     help='initial learning rate')
-parser.add_argument('--clip', type=float, default=None, help='gradient clipping')
-parser.add_argument('--bucket_type', type=str, default=None,
+parser.add_argument('--clip',
+                    type=float,
+                    default=None,
+                    help='gradient clipping')
+parser.add_argument('--bucket_type',
+                    type=str,
+                    default=None,
                     help='Can be "fixed" or "sorted"')
-parser.add_argument('--bucket_num', type=int, default=10,
+parser.add_argument('--bucket_num',
+                    type=int,
+                    default=10,
                     help='The bucket_num if bucket_type is "fixed".')
-parser.add_argument('--bucket_ratio', type=float, default=0.0,
+parser.add_argument('--bucket_ratio',
+                    type=float,
+                    default=0.0,
                     help='The ratio used in the FixedBucketSampler.')
-parser.add_argument('--bucket_mult', type=int, default=100,
+parser.add_argument('--bucket_mult',
+                    type=int,
+                    default=100,
                     help='The mult used in the SortedBucketSampler.')
-parser.add_argument('--valid_ratio', type=float, default=0.05,
-                    help='Proportion [0, 1] of training samples to use for validation set.')
-parser.add_argument('--epochs', type=int, default=20,
-                    help='upper epoch limit')
-parser.add_argument('--batch_size', type=int, default=16, metavar='N',
+parser.add_argument(
+    '--valid_ratio',
+    type=float,
+    default=0.05,
+    help='Proportion [0, 1] of training samples to use for validation set.')
+parser.add_argument('--epochs', type=int, default=3, help='upper epoch limit')
+parser.add_argument('--batch_size',
+                    type=int,
+                    default=16,
+                    metavar='N',
                     help='batch size')
-parser.add_argument('--dropout', type=float, default=0.,
+parser.add_argument('--dropout',
+                    type=float,
+                    default=0.,
                     help='dropout applied to layers (0 = no dropout)')
-parser.add_argument('--log-interval', type=int, default=30, metavar='N',
+parser.add_argument('--log-interval',
+                    type=int,
+                    default=30,
+                    metavar='N',
                     help='report interval')
-parser.add_argument('--save-prefix', type=str, default='sa-model',
+parser.add_argument('--save-prefix',
+                    type=str,
+                    default='sa-model',
                     help='path to save the final model')
-parser.add_argument('--gpu', type=int, default=None,
-                    help='id of the gpu to use. Set it to empty means to use cpu.')
+parser.add_argument(
+    '--gpu',
+    type=int,
+    default=None,
+    help='id of the gpu to use. Set it to empty means to use cpu.')
 args = parser.parse_args()
 print(args)
 
@@ -94,15 +130,17 @@ if args.gpu is None:
     context = mx.cpu()
 else:
     print('Use gpu%d' % args.gpu)
-    context = mx.gpu(args.gpu)
+    context = mx.gpu(int(args.gpu))
+
 
 class AggregationLayer(HybridBlock):
     """A block for different ways of aggregating encoder features"""
+
     def __init__(self, use_mean_pool=False, prefix=None, params=None):
         super(AggregationLayer, self).__init__(prefix=prefix, params=params)
         self._use_mean_pool = use_mean_pool
 
-    def hybrid_forward(self, F, data, valid_length): # pylint: disable=arguments-differ
+    def hybrid_forward(self, F, data, valid_length):  # pylint: disable=arguments-differ
         """Forward logic"""
         # Data will have shape (T, N, C)
         if self._use_mean_pool:
@@ -120,6 +158,7 @@ class AggregationLayer(HybridBlock):
 
 class SentimentNet(HybridBlock):
     """Network for sentiment analysis."""
+
     def __init__(self, dropout, use_mean_pool=False, prefix=None, params=None):
         super(SentimentNet, self).__init__(prefix=prefix, params=params)
         self._use_mean_pool = use_mean_pool
@@ -132,11 +171,12 @@ class SentimentNet(HybridBlock):
                 self.output.add(gluon.nn.Dropout(dropout))
                 self.output.add(gluon.nn.Dense(1, flatten=False))
 
-    def hybrid_forward(self, _, data, valid_length): # pylint: disable=arguments-differ
+    def hybrid_forward(self, _, data, valid_length):  # pylint: disable=arguments-differ
         encoded = self.encoder(self.embedding(data))  # Shape(T, N, C)
         agg_state = self.agg_layer(encoded, valid_length)
         out = self.output(agg_state)
         return out
+
 
 net = SentimentNet(dropout=args.dropout, use_mean_pool=args.use_mean_pool)
 with net.name_scope():
@@ -158,14 +198,20 @@ def preprocess(x):
     data = vocab[length_clip(tokenizer(data))]
     return data, label
 
+
 def get_length(x):
     return float(len(x[0]))
 
+
 # Load the dataset
-train_dataset, test_dataset = [nlp.data.IMDB(root='data/imdb', segment=segment)
-                               for segment in ('train', 'test')]
-train_dataset, valid_dataset = nlp.data.train_valid_split(train_dataset, args.valid_ratio)
+train_dataset, test_dataset = [
+    nlp.data.IMDB(root='data/imdb', segment=segment)
+    for segment in ('train', 'test')
+]
+train_dataset, valid_dataset = nlp.data.train_valid_split(
+    train_dataset, args.valid_ratio)
 print('Tokenize using spaCy...')
+
 
 def preprocess_dataset(dataset):
     start = time.time()
@@ -173,8 +219,10 @@ def preprocess_dataset(dataset):
     dataset = gluon.data.SimpleDataset(pool.map(preprocess, dataset))
     lengths = gluon.data.SimpleDataset(pool.map(get_length, dataset))
     end = time.time()
-    print('Done! Tokenizing Time={:.2f}s, #Sentences={}'.format(end - start, len(dataset)))
+    print('Done! Tokenizing Time={:.2f}s, #Sentences={}'.format(
+        end - start, len(dataset)))
     return dataset, lengths
+
 
 # Preprocess the dataset
 train_dataset, train_data_lengths = preprocess_dataset(train_dataset)
@@ -182,8 +230,9 @@ valid_dataset, valid_data_lengths = preprocess_dataset(valid_dataset)
 test_dataset, test_data_lengths = preprocess_dataset(test_dataset)
 
 # Construct the DataLoader. Pad data and stack label
-batchify_fn = nlp.data.batchify.Tuple(nlp.data.batchify.Pad(axis=0, pad_val=0, ret_length=True),
-                                      nlp.data.batchify.Stack(dtype='float32'))
+batchify_fn = nlp.data.batchify.Tuple(
+    nlp.data.batchify.Pad(axis=0, pad_val=0, ret_length=True),
+    nlp.data.batchify.Stack(dtype='float32'))
 if args.bucket_type is None:
     print('Bucketing strategy is not used!')
     train_dataloader = DataLoader(dataset=train_dataset,
@@ -193,29 +242,32 @@ if args.bucket_type is None:
 else:
     if args.bucket_type == 'fixed':
         print('Use FixedBucketSampler')
-        batch_sampler = nlp.data.FixedBucketSampler(train_data_lengths,
-                                                    batch_size=args.batch_size,
-                                                    num_buckets=args.bucket_num,
-                                                    ratio=args.bucket_ratio,
-                                                    shuffle=True)
+        batch_sampler = nlp.data.FixedBucketSampler(
+            train_data_lengths,
+            batch_size=args.batch_size,
+            num_buckets=args.bucket_num,
+            ratio=args.bucket_ratio,
+            shuffle=True)
         print(batch_sampler.stats())
     elif args.bucket_type == 'sorted':
         print('Use SortedBucketSampler')
-        batch_sampler = nlp.data.SortedBucketSampler(train_data_lengths,
-                                                     batch_size=args.batch_size,
-                                                     mult=args.bucket_mult,
-                                                     shuffle=True)
+        batch_sampler = nlp.data.SortedBucketSampler(
+            train_data_lengths,
+            batch_size=args.batch_size,
+            mult=args.bucket_mult,
+            shuffle=True)
     else:
         raise NotImplementedError
     train_dataloader = DataLoader(dataset=train_dataset,
                                   batch_sampler=batch_sampler,
                                   batchify_fn=batchify_fn)
 
-valid_dataloader = DataLoader(dataset=valid_dataset,
-                              batch_size=args.batch_size,
-                              shuffle=False,
-                              sampler=nlp.data.SortedSampler(valid_data_lengths),
-                              batchify_fn=batchify_fn)
+valid_dataloader = DataLoader(
+    dataset=valid_dataset,
+    batch_size=args.batch_size,
+    shuffle=False,
+    sampler=nlp.data.SortedSampler(valid_data_lengths),
+    batchify_fn=batchify_fn)
 
 test_dataloader = DataLoader(dataset=test_dataset,
                              batch_size=args.batch_size,
@@ -223,122 +275,98 @@ test_dataloader = DataLoader(dataset=test_dataset,
                              sampler=nlp.data.SortedSampler(test_data_lengths),
                              batchify_fn=batchify_fn)
 
-
 net.hybridize()
 print(net)
 if args.no_pretrained:
     net.initialize(mx.init.Xavier(), ctx=context)
 else:
     net.output.initialize(mx.init.Xavier(), ctx=context)
-trainer = gluon.Trainer(net.collect_params(), 'ftml', {'learning_rate': args.lr})
-loss = gluon.loss.SigmoidBCELoss()
+trainer = gluon.Trainer(net.collect_params(), 'ftml',
+                        {'learning_rate': args.lr})
+loss_fn = gluon.loss.SigmoidBCELoss()
 
 
-def evaluate(dataloader):
-    """Evaluate network on the specified dataset"""
-    total_L = 0.0
-    total_sample_num = 0
-    total_correct_num = 0
-    start_log_interval_time = time.time()
-    print('Begin Testing...')
-    for i, ((data, valid_length), label) in enumerate(dataloader):
-        data = mx.nd.transpose(data.as_in_context(context))
-        valid_length = valid_length.as_in_context(context).astype(np.float32)
-        label = label.as_in_context(context)
-        output = net(data, valid_length)
-        L = loss(output, label)
-        pred = (output > 0.5).reshape((-1,))
-        total_L += L.sum().asscalar()
-        total_sample_num += label.shape[0]
-        total_correct_num += (pred == label).sum().asscalar()
-        if (i + 1) % args.log_interval == 0:
-            print('[Batch {}/{}] elapsed {:.2f} s'.format(
-                i + 1, len(dataloader), time.time() - start_log_interval_time))
-            start_log_interval_time = time.time()
-    avg_L = total_L / float(total_sample_num)
-    acc = total_correct_num / float(total_sample_num)
-    return avg_L, acc
+class SentimentAnalysisBatchProcessor(BatchProcessor):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def evaluate_batch(self, estimator, val_batch, batch_axis=0):
+        data = mx.nd.transpose(val_batch[0][0])
+        valid_length = val_batch[0][1].astype(np.float32)
+        label = val_batch[1]
+        data_list = gluon.utils.split_and_load(data, estimator.context)
+        valid_length_list = gluon.utils.split_and_load(valid_length,
+                                                       estimator.context)
+        label_list = gluon.utils.split_and_load(label, estimator.context)
+        pred = [
+            estimator.eval_net(data, valid_length)
+            for data, valid_length in zip(data_list, valid_length_list)
+        ]
+        loss = [
+            estimator.evaluation_loss(y_hat, y)
+            for y_hat, y in zip(pred, label_list)
+        ]
+        pre_labels = [(pred_pro > 0.5).reshape(-1, ) for pred_pro in pred]
+        return data_list, label_list, pre_labels, loss
+
+    def fit_batch(self, estimator, train_batch, batch_axis=0):
+        data = mx.nd.transpose(train_batch[0][0])
+        valid_length = train_batch[0][1].astype(np.float32)
+        label = train_batch[1]
+        data_list = gluon.utils.split_and_load(data, estimator.context)
+        valid_length_list = gluon.utils.split_and_load(valid_length,
+                                                       estimator.context)
+        label_list = gluon.utils.split_and_load(label, estimator.context)
+        with autograd.record():
+            pred = [
+                estimator.net(data, valid_length)
+                for data, valid_length in zip(data_list, valid_length_list)
+            ]
+            loss = [
+                estimator.loss(y_hat, y).mean()
+                for y_hat, y in zip(pred, label_list)
+            ]
+        for l in loss:
+            l.backward()
+        # Clip gradient
+        if args.clip is not None:
+            grads = [
+                p.grad(estimator.context)
+                for p in self.net.collect_params().values()
+            ]
+            gluon.utils.clip_global_norm(grads, args.clip)
+        pre_labels = [(pred_pro > 0.5).reshape(-1, ) for pred_pro in pred]
+        return data_list, label_list, pre_labels, loss
 
 
-def train():
-    """Training process"""
-    start_pipeline_time = time.time()
+# Metric to monitor
+acc_metric = metric.Accuracy()
 
-    # Training/Testing
-    best_valid_acc = 0
-    stop_early = 0
-    for epoch in range(args.epochs):
-        # Epoch training stats
-        start_epoch_time = time.time()
-        epoch_L = 0.0
-        epoch_sent_num = 0
-        epoch_wc = 0
-        # Log interval training stats
-        start_log_interval_time = time.time()
-        log_interval_wc = 0
-        log_interval_sent_num = 0
-        log_interval_L = 0.0
+earlystopping_handler = EarlyStoppingHandler(monitor=acc_metric,
+                                             min_delta=0,
+                                             patience=3,
+                                             mode='max')
 
-        for i, ((data, valid_length), label) in enumerate(train_dataloader):
-            data = mx.nd.transpose(data.as_in_context(context))
-            label = label.as_in_context(context)
-            valid_length = valid_length.as_in_context(context).astype(np.float32)
-            wc = valid_length.sum().asscalar()
-            log_interval_wc += wc
-            epoch_wc += wc
-            log_interval_sent_num += data.shape[1]
-            epoch_sent_num += data.shape[1]
-            with autograd.record():
-                output = net(data, valid_length)
-                L = loss(output, label).mean()
-            L.backward()
-            # Clip gradient
-            if args.clip is not None:
-                grads = [p.grad(context) for p in net.collect_params().values()]
-                gluon.utils.clip_global_norm(grads, args.clip)
-            # Update parameter
-            trainer.step(1)
-            log_interval_L += L.asscalar()
-            epoch_L += L.asscalar()
-            if (i + 1) % args.log_interval == 0:
-                print('[Epoch %d Batch %d/%d] avg loss %g, throughput %gK wps' % (
-                    epoch, i + 1, len(train_dataloader),
-                    log_interval_L / log_interval_sent_num,
-                    log_interval_wc / 1000 / (time.time() - start_log_interval_time)))
-                # Clear log interval training stats
-                start_log_interval_time = time.time()
-                log_interval_wc = 0
-                log_interval_sent_num = 0
-                log_interval_L = 0
-        end_epoch_time = time.time()
-        valid_avg_L, valid_acc = evaluate(valid_dataloader)
-        test_avg_L, test_acc = evaluate(test_dataloader)
-        print('[Epoch %d] train avg loss %g, '
-              'valid acc %.4f, valid avg loss %g, '
-              'test acc %.4f, test avg loss %g, throughput %gK wps' % (
-                  epoch, epoch_L / epoch_sent_num,
-                  valid_acc, valid_avg_L, test_acc, test_avg_L,
-                  epoch_wc / 1000 / (end_epoch_time - start_epoch_time)))
+checkpoint_handler = CheckpointHandler(
+    model_dir='./',
+    model_prefix=args.save_prefix,
+    monitor=acc_metric,  # Monitors a metric
+    save_best=True,
+    mode='max')  # Save the best model in terms of
 
-        if valid_acc < best_valid_acc:
-            print('No Improvement.')
-            stop_early += 1
-            if stop_early == 3:
-                break
-        else:
-            # Reset stop_early if the validation loss finds a new low value
-            print('Observed Improvement.')
-            stop_early = 0
-            net.save_parameters(args.save_prefix + '_{:04d}.params'.format(epoch))
-            best_valid_acc = valid_acc
+est = Estimator(net=net,
+                loss=loss_fn,
+                train_metrics=[mx.metric.Loss()],
+                val_metrics=[mx.metric.Accuracy()],
+                trainer=trainer,
+                context=context,
+                batch_processor=SentimentAnalysisBatchProcessor())
 
-    net.load_parameters(glob.glob(args.save_prefix+'_*.params')[-1], context)
-    valid_avg_L, valid_acc = evaluate(valid_dataloader)
-    test_avg_L, test_acc = evaluate(test_dataloader)
-    print('Best validation loss %g, validation acc %.4f'%(valid_avg_L, valid_acc))
-    print('Best test loss %g, test acc %.4f'%(test_avg_L, test_acc))
-    print('Total time cost %.2fs'%(time.time()-start_pipeline_time))
-
-
-if __name__ == '__main__':
-    train()
+with warnings.catch_warnings():
+    warnings.simplefilter("ignore")
+    # Magic line
+    est.fit(train_data=train_dataloader,
+            val_data=valid_dataloader,
+            epochs=args.epochs,
+            event_handlers=[checkpoint_handler,
+                            earlystopping_handler])  # Add the event handlers
